Re-aligning the tests to RX RTM
- Aligning tests with latest Rx changes.
- Tests seem to be passing after a few runs, must be some concurrency
thing, possibly NCrunch introducing some strangeness. Need to look into
the removal of the scheduler on the ctor, but from the Rx docs it looks
like 'Immediate' is the default anyway.
- A bit of solution re-organising.
- Removing XNA project infos from the XNA adapter - this project does
not use any XNA features and as such this is not required.
- Using nuget for the adapter and test projects.
- Removing the post-build on the RedBadger XPF project - it was breaking
NCrunch, need to confirm what this was for.
- Adding more NCrunch stuff to the .gitignore.